// src/api/generated/api.ts
export type LoanApplicationWorkflowStatus = typeof LoanApplicationWorkflowStatus[keyof typeof LoanApplicationWorkflowStatus];


export const LoanApplicationWorkflowStatus = {
  AWAITING_INTENT_CONFIRMATION: 'AWAITING_INTENT_CONFIRMATION',
  TERMS_SHAPING: 'TERMS_SHAPING',
  AWAITING_LOAN_APPLICATION: 'AWAITING_LOAN_APPLICATION',
  BLOCKED: 'BLOCKED',
  PENDING_BANK_AUTHORIZATION: 'PENDING_BANK_AUTHORIZATION',
  AWAITING_UNDERWRITING_REVIEW: 'AWAITING_UNDERWRITING_REVIEW',
  PREPARING_DOCUMENTS: 'PREPARING_DOCUMENTS',
  DOCS_PENDING_CUSTOMER_SIGNING: 'DOCS_PENDING_CUSTOMER_SIGNING',
  VALIDATING_DOCUMENTS: 'VALIDATING_DOCUMENTS',
  BANK_SIGNING: 'BANK_SIGNING',
  PROCESSING_DISBURSEMENT: 'PROCESSING_DISBURSEMENT',
  DISBURSED: 'DISBURSED',
  WITHDRAWN: 'WITHDRAWN',
  TIMED_OUT: 'TIMED_OUT',
  BACK_OFFICE_PROCESSING: 'BACK_OFFICE_PROCESSING',
  REFRESHING_DATA: 'REFRESHING_DATA',
  AWAITING_APPLICATION_REVISION: 'AWAITING_APPLICATION_REVISION',
} as const;